# frontend/openApi/tallyj.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "TallyJ4 API",
    "description": "Election management and vote tallying system API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Ballots/election/{electionGuid}": {
      "get": {
        "tags": [
          "Ballots"
        ],
        "summary": "Gets a paginated list of ballots for the specified election.",
        "parameters": [
          {
            "name": "electionGuid",
            "in": "path",
            "description": "The GUID of the election.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "The page number (starting from 1).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The number of items per page (1-200).",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoPaginatedResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoPaginatedResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoPaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ballots/{guid}": {
      "get": {
        "tags": [
          "Ballots"
        ],
        "summary": "Gets a specific ballot by its GUID.",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "description": "The GUID of the ballot.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Ballots"
        ],
        "summary": "Updates an existing ballot.",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "description": "The GUID of the ballot to update.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The updated ballot data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBallotDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBallotDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBallotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Ballots"
        ],
        "summary": "Deletes a ballot by its GUID.",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "description": "The GUID of the ballot to delete.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Ballots": {
      "post": {
        "tags": [
          "Ballots"
        ],
        "summary": "Creates a new ballot.",
        "requestBody": {
          "description": "The ballot creation data.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBallotDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBallotDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBallotDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallotDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Import/ballots/{electionGuid}": {
      "post": {
        "tags": [
          "Import"
        ],
        "parameters": [
          {
            "name": "electionGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportBallotsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportBallotsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportBallotsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Votes/ballot/{ballotGuid}": {
      "get": {
        "tags": [
          "Votes"
        ],
        "parameters": [
          {
            "name": "ballotGuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VoteDtoListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteDtoListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteDtoListApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BallotDto": {
        "type": "object",
        "properties": {
          "ballotGuid": {
            "type": "string",
            "format": "uuid"
          },
          "ballotCode": {
            "type": "string",
            "nullable": true
          },
          "locationGuid": {
            "type": "string",
            "format": "uuid"
          },
          "locationName": {
            "type": "string",
            "nullable": true
          },
          "ballotNumAtComputer": {
            "type": "integer",
            "format": "int32"
          },
          "computerCode": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "teller1": {
            "type": "string",
            "nullable": true
          },
          "teller2": {
            "type": "string",
            "nullable": true
          },
          "voteCount": {
            "type": "integer",
            "format": "int32"
          },
          "votes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BallotDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/BallotDto"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BallotDtoPaginatedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BallotDto"
            },
            "nullable": true
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "hasPreviousPage": {
            "type": "boolean",
            "readOnly": true
          },
          "hasNextPage": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateBallotDto": {
        "type": "object",
        "properties": {
          "electionGuid": {
            "type": "string",
            "format": "uuid"
          },
          "computerCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ImportBallotsRequest": {
        "type": "object",
        "properties": {
          "csvContent": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBallotDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "teller1": {
            "type": "string",
            "nullable": true
          },
          "teller2": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VoteDto": {
        "type": "object",
        "properties": {
          "ballotGuid": {
            "type": "string",
            "format": "uuid"
          },
          "positionOnBallot": {
            "type": "integer",
            "format": "int32"
          },
          "personGuid": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "personFullName": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "personCombinedInfo": {
            "type": "string",
            "nullable": true
          },
          "onlineVoteRaw": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "VoteDtoListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteDto"
            },
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter your token in the text input below.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}